Move helper fuctions in to helpers.go
